WIP Untangle service from ovn requirement
When OVN is not enabled (the fact will be extracted from a CR field
listing the necessary drivers called `ml2Drivers`),

- Configure the listed ml2Drivers as mechanism_drivers.
- Remove ovn-router from the list of service_plugins.
- Don't create OVN (metadata-agent, ovn-agent) secrets.
- Don't wait for OVN DBs (nb and sb) to generate neutron-api config file.
- Don't include [ovn] section either.

Note that the default case of an empty ml2Drivers field is functionally
equivalent to when ml2Drivers=[ovn].

On the other hand, if `ovn` is enabled with some other driver(s), then the
service will wait for OVN deployment as usual.

The added tests assume the presence of `openvswitch` ml2 driver in the
neutron-api image. The driver is part of the upstream neutron tree. I am not
aware of any immediate plans to remove it from there, and/or exclude its
installation from the tcib images. If/when this changes, we may have to update
test cases to use a different driver to stand for an "other" driver.

(An alternative would be using one of test drivers, e.g. logger, but the RPM
subpackage that contains these - python-neutron-tests - is not included in the
service tcib image.)

Related-JIRA: https://issues.redhat.com/browse/OSPRH-3680

TODOs:

- add kuttl test that will configure it for non-ovn driver
  (openvswitch) and check that the service started successfully